services/config/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pandas
openpyxl
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pydantic_settings